+++ DSCResources/WindowsClient/WindowsClient.schema.psm1
@@ -35,7 +35,7 @@ using module ..\..\PowerStig.psm1
         All STIG rule IDs of the specified type are collected in an array and passed to the Skip-Rule
         function. Each rule follows the same process as the SkipRule parameter.
 #>
-Configuration WindowsClient
+configuration WindowsClient
 {
     [CmdletBinding()]
     param
@@ -99,7 +99,7 @@ Configuration WindowsClient
     Import-DscResource -ModuleName WindowsDefenderDSC -ModuleVersion 1.0.0.0
     . "$resourcePath\windows.ProcessMitigation.ps1"
 
-    Import-DscResource -ModuleName GPRegistryPolicyDsc -ModuleVersion 1.0.0
+    Import-DscResource -ModuleName GPRegistryPolicyDsc -ModuleVersion 1.0.1
     Import-DscResource -ModuleName PSDscResources -ModuleVersion 2.10.0.0
     . "$resourcePath\windows.Script.skip.ps1"
     . "$resourcePath\windows.Service.ps1"