+++ DSCResources/WindowsServer/WindowsServer.schema.psm1
@@ -38,7 +38,7 @@ using module ..\..\PowerStig.psm1
         All STIG rule IDs of the specified type are collected in an array and passed to the Skip-Rule
         function. Each rule follows the same process as the SkipRule parameter.
 #>
-Configuration WindowsServer
+configuration WindowsServer
 {
     [CmdletBinding()]
     param
@@ -98,7 +98,7 @@ Configuration WindowsServer
     Import-DscResource -ModuleName AuditPolicyDsc -ModuleVersion 1.2.0.0
     . "$resourcePath\windows.AuditPolicySubcategory.ps1"
 
-    Import-DscResource -ModuleName GPRegistryPolicyDsc -ModuleVersion 1.0.0
+    Import-DscResource -ModuleName GPRegistryPolicyDsc -ModuleVersion 1.0.1
     Import-DscResource -ModuleName PSDscResources -ModuleVersion 2.10.0.0
     . "$resourcePath\windows.Registry.ps1"
     . "$resourcePath\windows.Script.skip.ps1"