+++ Tests/Integration/DSCResources/Common.integration.ps1
@@ -87,6 +87,18 @@ Describe ($title + " $($stig.StigVersion) mof output") {
             }
         }
 
+        Context 'Single Backward Compatibility Exception' {
+            It "Should compile the MOF with STIG exception $($backCompatException.Keys) without throwing" {
+                {& $technologyConfig @testParameterList -BackwardCompatibilityException $backCompatException} | Should -Not -Throw
+            }
+        }
+
+        Context 'Multiple Backward Compatibility Exceptions' {
+            It "Should compile the MOF with STIG exceptions $($backCompatExceptionMultiple.Keys) without throwing" {
+                {& $technologyConfig @testParameterList -BackwardCompatibilityException $backCompatExceptionMultiple} | Should -Not -Throw
+            }
+        }
+
         Context 'Single Rule' {
             It 'Should compile the MOF without throwing' {
                 {& $technologyConfig @testParameterList -SkipRule $skipRule } | Should -Not -Throw