+++ Tests/Integration/DSCResources/IisSite.integration.tests.ps1
@@ -42,6 +42,8 @@ try
         $exception = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 1
         $getRandomExceptionRuleParams.RuleType = 'WebAppPoolRule'
         $exceptionMultiple = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 2
+        $backCompatException = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 1 -BackwardCompatibility
+        $backCompatExceptionMultiple = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 2 -BackwardCompatibility
 
         . "$PSScriptRoot\Common.integration.ps1"
     }