+++ Tests/Integration/DSCResources/InternetExplorer.integration.tests.ps1
@@ -34,6 +34,8 @@ try
         }
         $exception = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 1
         $exceptionMultiple = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 2
+        $backCompatException = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 1 -BackwardCompatibility
+        $backCompatExceptionMultiple = Get-RandomExceptionRule @getRandomExceptionRuleParams -Count 2 -BackwardCompatibility
 
         . "$PSScriptRoot\Common.integration.ps1"
     }